Backport rust uprev fixes (#801)
* use flatten() when possible

* from_iter -> collect

* unnecessary ? operator

* small and likely backward-compatible fixes

* small fixes

* use more specific and descriptive error type

+++ consensus/service/src/byzantine_ledger/mod.rs
@@ -156,7 +156,7 @@ impl ByzantineLedger {
                         if !worker.tick() {
                             break;
                         }
-                        thread::sleep(Duration::from_millis(10 as u64));
+                        thread::sleep(Duration::from_millis(10));
                     })
                     .expect("failed spawning ByzantineLedger"),
             )
